Add FluentResult

+++ WebShop/WebShopApp/Services/ClothesService.cs
@@ -1,4 +1,5 @@
 using AutoMapper;
+using FluentResults;
 using WebShopApp.Models.RequestModels;
 using WebShopApp.Models.ResponseModels;
 using WebShopData.Models;
@@ -16,36 +17,58 @@ namespace WebShopApp.Services
             _repository = repository;
             _mapper = mapper;
         }
-        public async Task<IEnumerable<ClothesItemsResponse>> GetAllAsync()
+        public async Task<Result<IEnumerable<ClothesItemsResponse>>> GetAllAsync()
         {
             var items = await _repository.GetAllAsync();
-            return _mapper.Map<IEnumerable<ClothesItemsResponse>>(items);
+            if (!items.Any())
+            {
+                return Result.Fail<IEnumerable<ClothesItemsResponse>>("No orders found.");
+            }
+            return Result.Ok(_mapper.Map<IEnumerable<ClothesItemsResponse>>(items));
         }
 
-        public async Task<ClothesItemsResponse?> GetByIdAsync(Guid id)
+        public async Task<Result<ClothesItemsResponse?>> GetByIdAsync(Guid id)
         {
             var item = await _repository.GetByIdAsync(id);
-            return _mapper.Map<ClothesItemsResponse?>(item);
+            if (item == null)
+            {
+                return Result.Fail("Nema takvog artikla bree");
+            }
+            var response = _mapper.Map<ClothesItemsResponse?>(item);
+            return Result.Ok(response);
         }
 
-        public async Task<ClothesItem> AddAsync(ClothesItemRequest request)
+        public async Task<Result<ClothesItem>> AddAsync(ClothesItemRequest request)
         {
             var item = _mapper.Map<ClothesItem>(request);
             item.Id = Guid.NewGuid(); // Generisanje ID-a
             await _repository.AddAsync(item);
-            return item; // Vraćamo kreirani entitet
+            return Result.Ok(item); // Vraćamo kreirani entitet
         }
 
-        public async Task<bool> UpdateAsync(Guid id, ClothesItemRequest request)
+        public async Task<Result> UpdateAsync(Guid id, ClothesItemRequest request)
         {
-            var item = _mapper.Map<ClothesItem>(request);
-            item.Id = id; // Postavi ID jer ga DTO nema
-            return await _repository.UpdateAsync(item);
+            var item = await _repository.GetByIdAsync(id);
+            if (item == null)
+            {
+                return Result.Fail("Nema takvog artikla bree");
+            }
+
+            _mapper.Map(request, item);
+            await _repository.UpdateAsync(item);
+            return Result.Ok();
         }
 
-        public async Task<bool> DeleteAsync(Guid id)
+        public async Task<Result> DeleteAsync(Guid id)
         {
-            return await _repository.DeleteAsync(id);
+            var exists = await _repository.GetByIdAsync(id);
+            if (exists == null)
+            {
+                return Result.Fail("Nema takvog artikla bree");
+            }
+
+            await _repository.DeleteAsync(id);
+            return Result.Ok();
         }
     }
 }
