+++ WebShop/WebShopApp/Services/TypesService.cs
@@ -1,4 +1,5 @@
 using AutoMapper;
+using FluentResults;
 using WebShopApp.Models.RequestModels;
 using WebShopApp.Models.ResponseModels;
 using WebShopData.Models;
@@ -17,36 +18,58 @@ namespace WebShopApp.Services
             _mapper = mapper;
         }
 
-        public async Task<IEnumerable<ClothesTypesResponse>> GetAllAsync()
+        public async Task<Result<IEnumerable<ClothesTypesResponse>>> GetAllAsync()
         {
             var items = await _repository.GetAllAsync();
-            return _mapper.Map<IEnumerable<ClothesTypesResponse>>(items);
+            if (!items.Any())
+            {
+                return Result.Fail<IEnumerable<ClothesTypesResponse>>("No orders found.");
+            }
+            return Result.Ok(_mapper.Map<IEnumerable<ClothesTypesResponse>>(items));
         }
 
-        public async Task<ClothesTypesResponse?> GetByIdAsync(Guid id)
+        public async Task<Result<ClothesTypesResponse?>> GetByIdAsync(Guid id)
         {
             var item = await _repository.GetByIdAsync(id);
-            return _mapper.Map<ClothesTypesResponse?>(item);
+            if (item == null)
+            {
+                return Result.Fail("Item not found.");
+            }
+            var response = _mapper.Map<ClothesTypesResponse?>(item);
+            return Result.Ok(response);
         }
 
-        public async Task<ClothesType> AddAsync(ClothesTypeRequest typeRequest)
+        public async Task<Result<ClothesType>> AddAsync(ClothesTypeRequest typeRequest)
         {
             var type = _mapper.Map<ClothesType>(typeRequest);
             type.Id = Guid.NewGuid(); 
             await _repository.AddAsync(type);
-            return type;
+            return Result.Ok(type);
         }
 
-        public async Task<bool> UpdateAsync(Guid id, ClothesTypeRequest typeRequest)
+        public async Task<Result> UpdateAsync(Guid id, ClothesTypeRequest typeRequest)
         {
-            var type = _mapper.Map<ClothesType>(typeRequest);
-            type.Id = id;
-            return await _repository.UpdateAsync(type);
+            var item = await _repository.GetByIdAsync(id);
+            if (item == null)
+            {
+                return Result.Fail("Item not found.");
+            }
+
+            _mapper.Map(typeRequest, item);
+            await _repository.UpdateAsync(item);
+            return Result.Ok();
         }
 
-        public async Task<bool> DeleteAsync(Guid id)
+        public async Task<Result> DeleteAsync(Guid id)
         {
-            return await _repository.DeleteAsync(id);
+            var exists = await _repository.GetByIdAsync(id);
+            if (exists == null)
+            {
+                return Result.Fail("Item not found.");
+            }
+
+            await _repository.DeleteAsync(id);
+            return Result.Ok();
         }
     }
 }
